The files so far for the bare bones battleship
It's not finished yet but it's close. To run, downlaod the files, go to the files then right click "index.html" and select "Copy Full Path". Go to your browser and paste and the game will come up. 

While it isn't finished as of yet and you can't drag your peices to your side or play the game with the AI yet, it should be finished very soon